requirements.txt
flask 
numpy 
Pillow
